refactor: 변경된 카드 생성 로직 적용
