+++ src/main/java/blackjack/model/deck/DeckManager.java
@@ -1,4 +1,4 @@
-package blackjack.model.generator;
+package blackjack.model.deck;
 
 import blackjack.model.cards.Card;
 import java.util.ArrayList;
