--- src/main/java/blackjack/model/generator/IndexGenerator.java
@@ -1,6 +1,0 @@
-package blackjack.model.generator;
-
-@FunctionalInterface
-public interface IndexGenerator {
-    int generate(int maxRange);
-}
